+++ source/base/crypto/message_decryptor_openssl.cc
@@ -96,7 +96,8 @@ bool MessageDecryptorOpenssl::decrypt(const void* in, size_t in_size, void* out)
 
     if (EVP_DecryptUpdate(ctx_.get(),
                           reinterpret_cast<uint8_t*>(out), &length,
-                          reinterpret_cast<const uint8_t*>(in) + kTagSize, in_size - kTagSize) != 1)
+                          reinterpret_cast<const uint8_t*>(in) + kTagSize,
+                          static_cast<int>(in_size) - kTagSize) != 1)
     {
         LOG(LS_WARNING) << "EVP_DecryptUpdate failed";
         return false;
